Struct imgui_memory_editor::MemoryEditor

pub struct MemoryEditor { /* fields omitted */ }

Dear ImGui memory editor widget.

Use by calling either draw_contents or draw_window methods.

Note:

While theoretically this struct is Send + Sync compatible, it doesn't implement those traits, since Dear ImGui isn't thread safe.

Implementations

impl MemoryEditor

pub fn new() -> Self

Create a new MemoryEditor instance.

pub fn get_open(&self) -> bool

Get the Open field.

pub fn set_open(&mut self, open: bool)

Set the Open field.

pub fn get_read_only(&self) -> bool

Get the ReadOnly field.

pub fn set_read_only(&mut self, read_only: bool)

Set the ReadOnly field.

pub fn get_cols(&self) -> c_int

Get the Cols field.

pub fn set_cols(&mut self, columns: c_int)

Set the Cols field.

pub fn get_highlight_color(&self) -> ImColor

Get the HighlightColor field.

pub fn set_highlight_color(&mut self, color: ImColor)

Set the HighlightColor field.

pub fn get_show_hexii(&self) -> bool

Get the OptShowHexII field.

pub fn set_show_hexii(&mut self, show: bool)

Set the OptShowHexII field.

pub fn get_show_ascii(&self) -> bool

Get the OptShowAscii field.

pub fn set_show_ascii(&mut self, show: bool)

Set the OptShowAscii field.

pub fn get_show_options(&self) -> bool

Get the OptShowOptions field.

pub fn set_show_options(&mut self, show: bool)

Set the OptShowOptions field.

pub fn get_show_data_preview(&self) -> bool

Get the OptShowDataPreview field.

pub fn set_show_data_preview(&mut self, show: bool)

Set the OptShowDataPreview field.

pub fn get_grey_out_zeroes(&self) -> bool

Get the OptGreyOutZeroes field.

pub fn set_grey_out_zeroes(&mut self, greyout: bool)

Set the OptGreyOutZeroes field.

pub fn get_upper_case_hex(&self) -> bool

Get the OptUpperCaseHex field.

pub fn set_upper_case_hex(&mut self, uppercase: bool)

Set the OptUpperCaseHex field.

pub fn get_mid_cols_count(&self) -> c_int

Get the OptMidColsCount field.

pub fn set_mid_cols_count(&mut self, count: c_int)

Set the OptMidColsCount field.

pub fn get_addr_digits_count(&self) -> c_int

Get the OptAddrDigitsCount field.

pub fn set_addr_digits_count(&mut self, count: c_int)

Set the OptAddrDigitsCount field.

pub fn set_read_fn<F>(&mut self, _: F) where
F: Fn(*const ImU8, usize) -> ImU8,

Set the ReadFn field.

You can only pass in ordinary functions and stateless closures, if you pass anything else this function will panic.

pub fn set_write_fn<F>(&mut self, _: F) where
F: Fn(*mut ImU8, usize, ImU8),

Set the WriteFn field.

You can only pass in ordinary functions and stateless closures, if you pass anything else this function will panic.

pub fn set_highlight_fn<F>(&mut self, _: F) where
F: Fn(*const ImU8, usize) -> bool,

Set the HighlightFn field.

You can only pass in ordinary functions and stateless closures, if you pass anything else this function will panic.

pub fn draw_contents(
&mut self,
_: &Ui<'_>,
mem_data: &mut [u8],
base_display_addr: Option<usize>
)

Render only the contents of the memory editor, without any window.

The base_display_addr field has a default value of 0x0000 if None is passed as an argument.

pub fn draw_window(
&mut self,
_: &Ui<'_>,
title: &ImStr,
mem_data: &mut [u8],
base_display_addr: Option<usize>
)

Render standalone memory editor, in a window.

The base_display_addr field has a default value of 0x0000 if None is passed as an argument.